Moved the test and changed the comment of that test
