--- core/src/test/java/org/geysermc/geyser/scoreboard/network/sidebar/ReloadSidebarScoreboardTest.java
@@ -1,110 +1,0 @@
-/*
- * Copyright (c) 2026 GeyserMC. http://geysermc.org
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- *
- * @author GeyserMC
- * @link https://github.com/GeyserMC/Geyser
- */
-
-package org.geysermc.geyser.scoreboard.network.sidebar;
-
-import static org.geysermc.geyser.scoreboard.network.util.AssertUtils.assertNextPacket;
-import static org.geysermc.geyser.scoreboard.network.util.AssertUtils.assertNoNextPacket;
-import static org.geysermc.geyser.scoreboard.network.util.GeyserMockContextScoreboard.mockContextScoreboard;
-
-import java.util.List;
-import net.kyori.adventure.text.Component;
-import org.cloudburstmc.protocol.bedrock.data.ScoreInfo;
-import org.cloudburstmc.protocol.bedrock.packet.SetDisplayObjectivePacket;
-import org.cloudburstmc.protocol.bedrock.packet.SetScorePacket;
-import org.geysermc.geyser.translator.protocol.java.scoreboard.JavaResetScorePacket;
-import org.geysermc.geyser.translator.protocol.java.scoreboard.JavaSetDisplayObjectiveTranslator;
-import org.geysermc.geyser.translator.protocol.java.scoreboard.JavaSetObjectiveTranslator;
-import org.geysermc.geyser.translator.protocol.java.scoreboard.JavaSetScoreTranslator;
-import org.geysermc.mcprotocollib.protocol.data.game.scoreboard.ObjectiveAction;
-import org.geysermc.mcprotocollib.protocol.data.game.scoreboard.ScoreType;
-import org.geysermc.mcprotocollib.protocol.data.game.scoreboard.ScoreboardPosition;
-import org.geysermc.mcprotocollib.protocol.packet.ingame.clientbound.scoreboard.ClientboundResetScorePacket;
-import org.geysermc.mcprotocollib.protocol.packet.ingame.clientbound.scoreboard.ClientboundSetDisplayObjectivePacket;
-import org.geysermc.mcprotocollib.protocol.packet.ingame.clientbound.scoreboard.ClientboundSetObjectivePacket;
-import org.geysermc.mcprotocollib.protocol.packet.ingame.clientbound.scoreboard.ClientboundSetScorePacket;
-import org.junit.jupiter.api.Test;
-
-/**
- * Test for <a href="https://github.com/GeyserMC/Geyser/issues/6609">#6609</a>.
- *
- * <p>A scoreboard plugin reload resets every score and then re-adds it with identical text and
- * values. The re-added score has to reach Bedrock again, because the reset already removed it
- * there - a leftover entry in the render cache used to swallow it, leaving the line permanently
- * missing until its text changed.</p>
- */
-public class ReloadSidebarScoreboardTest {
-
-    @Test
-    void resetThenReaddSameLine() {
-        mockContextScoreboard(context -> {
-            var setObjectiveTranslator = new JavaSetObjectiveTranslator();
-            var setDisplayObjectiveTranslator = new JavaSetDisplayObjectiveTranslator();
-            var setScoreTranslator = new JavaSetScoreTranslator();
-            var resetScoreTranslator = new JavaResetScorePacket();
-
-            context.translate(setObjectiveTranslator, new ClientboundSetObjectivePacket(
-                "objective", ObjectiveAction.ADD, Component.text("objective"), ScoreType.INTEGER, null));
-            context.translate(setDisplayObjectiveTranslator,
-                new ClientboundSetDisplayObjectivePacket(ScoreboardPosition.SIDEBAR, "objective"));
-            assertNextPacket(context, () -> {
-                var packet = new SetDisplayObjectivePacket();
-                packet.setObjectiveId("0");
-                packet.setDisplayName("objective");
-                packet.setCriteria("dummy");
-                packet.setDisplaySlot("sidebar");
-                packet.setSortOrder(1);
-                return packet;
-            });
-
-            context.translate(setScoreTranslator, new ClientboundSetScorePacket("line1", "objective", 1));
-            assertNextPacket(context, () -> {
-                var packet = new SetScorePacket();
-                packet.setAction(SetScorePacket.Action.SET);
-                packet.setInfos(List.of(new ScoreInfo(1, "0", 1, "line1")));
-                return packet;
-            });
-
-            // the plugin clears its board
-            context.translate(resetScoreTranslator, new ClientboundResetScorePacket("line1", "objective"));
-            assertNextPacket(context, () -> {
-                var packet = new SetScorePacket();
-                packet.setAction(SetScorePacket.Action.REMOVE);
-                packet.setInfos(List.of(new ScoreInfo(1, "0", 0)));
-                return packet;
-            });
-
-            // and re-adds the identical line, which must be sent to Bedrock again
-            context.translate(setScoreTranslator, new ClientboundSetScorePacket("line1", "objective", 1));
-            assertNextPacket(context, () -> {
-                var packet = new SetScorePacket();
-                packet.setAction(SetScorePacket.Action.SET);
-                packet.setInfos(List.of(new ScoreInfo(2, "0", 1, "line1")));
-                return packet;
-            });
-            assertNoNextPacket(context);
-        });
-    }
-}
